% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caesarian.R
\docType{data}
\name{caesarian}
\alias{caesarian}
\title{Caesarian Classification}
\format{
A data frame with 80 rows and 6 variables:
\describe{
\item{\code{age}}{integer; Age in years.}
\item{\code{delivery.num}}{integer; Delivery number.}
\item{\code{delivery.time}}{factor; \code{timely, premature, latecomer}.}
\item{\code{bp}}{factor; \code{low, normal, high}.}
\item{\code{heart.problem}}{factor; \code{apt, inept}.}
\item{\code{caesarian}}{factor; \code{no, yes}.}
}
}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/Caesarian+Section+Classification+Dataset}{ICU Machine Learning Repository}.

Provided by
Muhammad Zain Amin, University of Engineering and Technology, Lahore, Pakistan
Amir Ali, University of Engineering and Technology, Lahore, Pakistan
}
\usage{
caesarian
}
\description{
Predicting caesarian births from medical and demographic data.
}
\references{
M.Zain Amin, Amir Ali.'Performance Evaluation of Supervised
Machine Learning Classifiers for Predicting Healthcare Operational Decisions'.
Machine Learning for Operational Decision Making,
Wavy Artificial Intelligence Research Foundation, Pakistan, 2018
}
\keyword{datasets}
