% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_plots.R
\name{cr_plots}
\alias{cr_plots}
\title{Component+Residual Plots}
\usage{
cr_plots(x, alpha = 0.4, span = 0.75)
}
\arguments{
\item{x}{an object of type \code{"lm"}.}

\item{alpha}{numeric; degree of transparency for points (0 to 1, default=0.4)}

\item{span}{numeric; the  degree of smoothing for loess lines (default=0.75)}
}
\value{
a \code{ggplot2} graph
}
\description{
Produces component+residual (partial residual) plots for
linear models.
}
\details{
This function is a modification of the \link[car]{crPlots} function
in the \code{car} package, using \code{ggplot2} rather than
\code{base} graphics.
}
\note{
The linear model can contain factors but not interactions. Plots are
based on all predictors, but only the numeric variables are plotted. Each plot
has a linear and loess fit line.
}
\examples{
mtcars$am <- factor(mtcars$am)
fit <- lm(mpg ~ wt + am + disp + hp, mtcars)
cr_plots(fit)

}
\seealso{
\link[car]{crPlot}, \link[car]{crPlots}
}
