% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.glm.R
\name{diagnostics.glm}
\alias{diagnostics.glm}
\title{Diagnostics for Logistic Regression}
\usage{
\method{diagnostics}{glm}(x, alpha = 0.4, span = 0.8, ...)
}
\arguments{
\item{x}{an object of class \code{c("glm")}}

\item{alpha}{numeric; transparency for plot points (default=0.4)}

\item{span}{numeric; smoothing parameter for loess fit lines (default=0.8)}

\item{...}{not currently used}
}
\value{
a list containing the \code{ggplot2} graphs and statistical tests
}
\description{
Provides diagnostics tests and graphs for logistic regression
}
\details{
The \code{diagnostics} function is a wrapper for several
diagnostic graphing functions and tests. The function evaluates
the following:
\describe{
\item{Linearity}{Linearity of the explanatory-response relationship
are assessed via Ceres plots}
(\link[car]{ceresPlots}).
\item{Multicollinearity}{Variance inflation factors (\link[car]{vif})
are plotted. If there is a single predictor variable, this section
is skipped.}
\item{Outliers, leverage, and influence}{An influence plot
(\link{gginfluencePlot}) and Bonferroni Outlier Test
(\link[car]{outlierTest}) are provided.}
}
}
\examples{
mtcars$cyl <- factor(mtcars$cyl)
fit <- glm(am ~ hp + cyl + wt, family=binomial, mtcars)
diagnostics(fit)
}
\seealso{
\code{\link{diagnostics}}, \code{\link[car]{vif}},
\code{\link[car]{outlierTest}}, \code{\link[car]{influencePlot}},
\code{\link[car]{ceresPlots}}
}
