% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.lm.R
\name{diagnostics.lm}
\alias{diagnostics.lm}
\title{Regression Diagnostics for Linear Models}
\usage{
\method{diagnostics}{lm}(x, alpha = 0.4, span = 0.8, plot = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{"lm"}}

\item{alpha}{numeric; transparency for plot points (default=0.4)}

\item{span}{numeric; smoothing parameter for loess fit lines (default=0.8)}

\item{plot}{logical; If \code{TRUE} (the default), graphs are printed. Otherwise,
they are returned invisibly.}

\item{...}{not currently used}
}
\value{
A five component list containing \code{ggplot2} graphs:
qqplot, crplots, slplot, vifplot, and influenceplot.
}
\description{
Provides regression diagnostics for a linear models fit with
\code{\link{lm}} or \code{\link{regress}}.
}
\details{
The \code{diagnostics} function is a wrapper for several
diagnostic plotting functions:
\describe{
\item{Normality}{Normality of the (studentized) residuals is assessed
via a Normal Q-Q plot (\code{\link{qq_plot}}).}
\item{Linearity}{Linearity of the explanatory-response relationships
are assessed via Component + Residual (partial residual) plots
(\code{\link{cr_plots}}). If there is a single predictor, a scatter plot
with linear and loess lines is produced.}
\item{Constant variance}{Homoscedasticity is evaluated via
a Spread-Level plot (\code{\link{spread_plot}}).}
\item{Multicollinearity}{Multicollinearity is assessed via variance inflation factors
(\code{\link{vif_plot}}). If there is a single predictor variable, this section
is skipped.}
\item{Outliers, leverage, and influence}{A influence plot identifies
outliers and influential observations (\code{\link{influence_plot}}).}
}
}
\note{
Each function relies heavily on the \code{car} package. See the
help for individual functions for details.
}
\examples{
fit <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
diagnostics(fit)
}
\seealso{
\code{\link{diagnostics}}, \code{\link[car]{vif}},
\code{\link[car]{qqPlot}}, \code{\link[car]{outlierTest}}, \code{\link[car]{influencePlot}},
\code{\link[car]{crPlots}}, \code{\link[car]{spreadLevelPlot}}, \code{\link[car]{ncvTest}}
}
