% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.lreg.R
\name{diagnostics.lreg}
\alias{diagnostics.lreg}
\title{Diagnostics for Logistic Regression}
\usage{
\method{diagnostics}{lreg}(x)
}
\arguments{
\item{x}{an object of class \code{lreg}}
}
\description{
Provides diagnostics tests and graphs for a logistic regression model
}
\details{
Provides diagnostics including a test of
multicollinearity (\code{\link[car]{vif}}),
a graph for evaluating linearity (\code{\link[car]{ceresPlots}}),
a test for identifying outliers (\code{\link[car]{outlierTest}}),
and a graph for evaluating influential observations
(\code{\link[car]{influencePlot}}).
}
\examples{
mtcars$am <- factor(mtcars$am)
fit <- lreg(am ~ hp + wt + mpg, mtcars)
diagnostics(fit)
}
\seealso{
\code{\link[qacReg]{diagnostics}}, \code{\link[car]{vif}},
\code{\link[car]{ceresPlots}}, \code{\link[car]{outlierTest}}, \code{\link[car]{influencePlot}}
}
