% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{diagnostics:

Run and print regression diagnostics}
\usage{
diagnostics(x, ...)
}
\arguments{
\item{x}{an object of type \code{mreg} or \code{lreg}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The results of the methods \code{\link{diagnostics.mreg}} or \code{\link{diagnostics.lreg}}
}
\description{
diagnostics:

Run and print regression diagnostics
}
\examples{
fit <- mreg(mpg ~ hp + wt + accel + origin, data = auto_mpg)
diagnostics(fit)
}
