% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggforestPlot.R
\name{ggforestPlot}
\alias{ggforestPlot}
\title{Forest Plot}
\usage{
ggforestPlot(x, digits = 2, ci = 0.95)
}
\arguments{
\item{x}{object of class "lm" or "glm"}

\item{digits}{number of decimal digits to display}

\item{ci}{confidence level (default = 0.95)}
}
\value{
a ggplot2 graph
}
\description{
Display a forest plot of regression parameters for a
given \code{"lm"} or \code{"glm"} model.
}
\details{
For linear models, the plot displays the standardized coefficients and
their confidence intervals. For logistic regression models, the plot
displays odds ratios and their confidence intervals. Effects are plotted
in the same order as they appear in the model.
}
\examples{
#######################
# multiple regression #
#######################
fit <- lm(mpg ~ ., data = auto_mpg)
ggforestPlot(fit)

#######################
# logistic regression #
#######################
fit2 <- glm(caesarian ~ ., family = binomial, data = caesarian)
ggforestPlot(fit2)
}
\seealso{
\code{\link{info.lm}}, \code{\link{info.glm}}
}
