% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{ggliftPlot}
\alias{ggliftPlot}
\title{Lift and gain charts}
\usage{
ggliftPlot(actual, prob, positive = "yes")
}
\arguments{
\item{actual}{actual class.}

\item{prob}{predicted probability of target class.}

\item{positive}{which class is the target class?}
}
\value{
a data frame with lift and gain values for each decile.
}
\description{
Create lift and gain charts for binary
classification problems
}
\examples{
# logistic regression example
fit <- glm(am ~ . , mtcars, family=binomial)
pred <- predict(fit, newdata=mtcars, type="response")
results <- ggliftPlot(mtcars$am, pred, positive="1")
print(results)
}
