% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggqqPlot.R
\name{ggqqPlot}
\alias{ggqqPlot}
\title{Normal Q-Q Plot}
\usage{
ggqqPlot(x, reps = 100, conf = 0.95, n.labels = 3)
}
\arguments{
\item{x}{an object of type \code{"mreg"} or \code{"lm"}.}

\item{reps}{integer; number of bootstrap replications for
confidence envelope.}

\item{conf}{size of confidence interval (default=\code{0.95}).}

\item{n.labels}{integer; the number of largest residuals to label.}
}
\description{
Plots empirical quantiles of the studentized residuals
from a linear model against the theoretical quantiles of
a the normal distribution.
}
\details{
This function is a modification of the \link[car]{qqPlot} function
in the \code{car} package, using \code{ggplot3} rather than
\code{base} graphics.
}
\seealso{
\link[car]{qqPlot}, \link[stats]{qqnorm}
}
