% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspreadlevel.R
\name{ggSpreadLevelPlot}
\alias{ggSpreadLevelPlot}
\title{Spread-Level Plot}
\usage{
ggSpreadLevelPlot(x, n.labels = 3)
}
\arguments{
\item{x}{an object of type \code{"mreg"} or \code{"lm"}.}

\item{n.labels}{integer; the number of largest residuals to label.}
}
\description{
Creates a plot for examining the possible dependence of
spread on level using the studentized residuals from a linear
model.
}
\details{
This function is a modification of the \link[car]{spreadLevelPlot} function
in the \code{car} package, using \code{ggplot2} rather than
\code{base} graphics.
}
\note{
The graph plots the log of the fitted values against the log
of the absolute studentized residuals. A robust linear fit line
and a loess fit line are also plotted.
}
\seealso{
\link[car]{spreadLevelPlot}
}
