% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvif.R
\name{ggvif}
\alias{ggvif}
\title{Variance Inflation Plot}
\usage{
ggvif(x)
}
\arguments{
\item{x}{an object of type \code{"mreg"} or \code{"lm"}.}
}
\description{
Plots the variance inflations factors (VIF) for a linear
model.
}
\details{
This function plots results from the \link[car]{vif} function.
If all terms in an unweighted linear model have 1 df, then the usual variance-inflation factors are calculated.
If any terms in an unweighted linear model have more than 1 df,
then generalized variance-inflation factors (Fox and Monette, 1992)
are calculated.
}
\references{
Fox, J. and Monette, G. (1992) Generalized collinearity diagnostics.
JASA, 87, 178–183.
}
\seealso{
\link[car]{vif}
}
