% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.glm.R
\name{info.glm}
\alias{info.glm}
\title{Summarise a logistic regression model}
\usage{
\method{info}{glm}(x)
}
\arguments{
\item{x}{an object of class "glm"}
}
\value{
The list of class c(\code{"info.glm"}, \code{"list"}).
\describe{
\item{omnibus.test}{omnibus chisquare test}
\item{fit.indices}{goodness of fit measures including Stukel's test, Pseudo R-squared, and Akaike Information Criterion}
\item{anova.table}{Type-III Analysis of Variance Table}
\item{coefficient.table}{includes Coefficients, Standardized Coefficients,
Standard Errors, z-values, and p-values with significant stars}
\item{oddsratios}{odd's ratios with 95\% confidence intervals}
\item{N}{sample size}
\item{target}{category predicted}
}
}
\description{
Summarize a logistic regression model of class \code{"glm"}.
}
\details{
Put details here
}
\examples{
fit <- regress(am ~ hp + wt + mpg, mtcars)
info(fit)
}
\seealso{
\code{\link[car]{Anova}}, \code{\link[car]{Confint}}
}
