% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.lm.R
\name{info.lm}
\alias{info.lm}
\title{Model Summary for an 'lm' Object}
\usage{
\method{info}{lm}(x)
}
\arguments{
\item{x}{an object of class \code{"lm"}}
}
\value{
A list of class c(\code{"info.lm"}, \code{"list"}).
\itemize{
\item{overall}
\item{fit.indices}
\item{F.test}
\item{anova.table}
\item{coefficient.table}
}
}
\description{
Extended summary information for an object of class \code{"lm"}.
}
\details{
The function \code{info.lm} produces a summary for a
linear model fitted with the \code{\link{lm}}  or
\code{\link{regress}} function. There are five sections.

\describe{
\item{Overall}{Model formula, data frame, and sample size (N).}
\item{Fit Indices}{R squared, adjusted R squared, Akaike's information
criterion (AIC), root mean square error (RMSE), and
mean absolute deviation (MAE).}
\item{Omnibus Test}{F-statistic, degrees of freedom, p-value.}
\item{Analysis of Variance}{ANOVA table with type III (marginal)
effects.}
\item{Regression Coefficients}{Regression coefficients (B),
standardized regression coefficients (B*) standard errors (SE),
t-values, and p-values.}
}

The ANOVA table is obtained from the \code{\link[car]{Anova}} function in
the \code{car} package. The standardized regression coefficients are
obtained for the \code{\link[lm.beta]{lm.beta}} function in the
\code{lm.beta} package.
}
\examples{
fit <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
info(fit)
}
\seealso{
\code{\link{print.info.lm}}
}
