% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.glm.R
\name{performance.glm}
\alias{performance.glm}
\title{Model Performance for Logistic Regression}
\usage{
\method{performance}{glm}(x, data, digits = 4, p = 0.5, plot = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{"glm"}.}

\item{data}{a data frame. If no dataset is provided, the model training
data is used.}

\item{digits}{integer; number of digits to print (default=4).}

\item{p}{numeric; probability cutoff for classifying cases (default = 0.5)}

\item{plot}{logical; If \code{TRUE} the results are plotted.}

\item{...}{not currently used}
}
\value{
The results of the methods \code{\link{performance.glm}}
}
\description{
Model Performance for Logistic Regression
}
\details{
Produce indices of model performance for logistic regression
}
\examples{
# performance on training sample
fit <- glm(am ~ hp + wt, family=binomial, data = mtcars)
performance(fit)
}
