% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.lm.R
\name{performance.lm}
\alias{performance.lm}
\title{Model Performance for Linear Models}
\usage{
\method{performance}{lm}(x, data, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{"lm"}.}

\item{data}{a data frame.}

\item{digits}{integer; number of digits to print (default=4).}

\item{...}{not currently used}
}
\value{
A vector of performance estimates (RMSE, Rsquared, and MAE).
}
\description{
Produce indices of model performance for linear models
}
\details{
\code{performance} calculates the r-squared, root mean square error (RMSE),
and mean absolute error (MAE) applying the model to a data frame. If
a data frame is not specified, the model is evaluated on the
training data (resubstitution). Results are provided by the
\code{\link[caret]{postResample}} function.
}
\examples{
# performance on training sample
fit <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
performance(fit)
}
\seealso{
\code{\link[caret]{postResample}}
}
