% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Model Performance}
\usage{
performance(x, ...)
}
\arguments{
\item{x}{an object of type \code{"lm"} or \code{"glm"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The results of the methods \code{\link{performance.lm}} or
\code{\link{performance.glm}}
}
\description{
Indices of model performance for linear and logistic regression
models.
}
\details{
details go here.
}
\examples{
#######################
# multiple regression #
#######################
fit <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
performance(fit)

#######################
# logistic regression #
#######################
fit2 <- glm(caesarian ~ age + bp + delivery.time, family = binomial, data = caesarian)
performance(fit2)

}
