% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mreg.R
\name{plot.mreg}
\alias{plot.mreg}
\title{Plotting multiple regression}
\usage{
\method{plot}{mreg}(x, points = FALSE)
}
\arguments{
\item{x}{an object of type \code{c("mreg", "lm")}}

\item{points}{points is TRUE if data points are to be shown, default is FALSE}
}
\value{
a plot of class \code{ggplot2}
}
\description{
The plot function provides a single plot that shows separate graphs of the relationship
between each predictor and the outcome, controlling for the other predictors from the
mreg function
}
\details{
The \code{plot} function is uses ggeffects to plot the results of the mreg function.
The ggeffects package aims at easily calculating marginal effects and adjusted predictions
at representative values of covariates from statistical models,
i.e. predictions generated by a model when one holds the controlled dependent variables constant and varies the independent variable(s).
To see more details about the ggeffects package for better understanding see
\link[https://strengejacke.github.io/ggeffects/]{here}
}
\examples{
mtcars$am <- factor(mtcars$am)
fit <- mreg(mpg ~ hp + wt + am, mtcars)
plot(fit)
plot(fit, points=TRUE)
}
\seealso{
\link[ggeffects]{ggeffect}
\link[ggplot2]{ggplot}
}
