% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.info.lm.R
\name{print.info.lm}
\alias{print.info.lm}
\title{Print "info.lm" Object}
\usage{
\method{print}{info.lm}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class \code{"info.lm"} produced by the
\code{"info.lm"} function.}

\item{digits}{number of significant digits to print.}

\item{...}{parameters passed to print.}
}
\description{
\code{print} method for class \code{"info.lm"}.
}
\examples{
fit <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
sfit <- info(fit)
print(sfit, digits=5)
}
