% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.mreg.R
\name{print.summary.mreg}
\alias{print.summary.mreg}
\title{Print Summary of a Linear Model Fit}
\usage{
\method{print}{summary.mreg}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class \code{"summary.mreg"} produced by the
\code{"summary.mreg"} function.}

\item{digits}{number of significant digits to print.}

\item{...}{parameters passed to print.}
}
\description{
\code{print} method for class \code{"summary.mreg"}.
}
\examples{
fit <- mreg(mpg ~ wt + hp + disp, data=mtcars)
sfit <- summary(fit)
print(sfit)
}
