% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{Regression Analysis}
\usage{
regress(formula, data, ...)
}
\arguments{
\item{formula}{on object of class \code{"formula"}}

\item{data}{a data frame}

\item{...}{optional parameters pass to the fitting functions}
}
\value{
an object of class \code{"lm"} or \code{"glm"}.
}
\description{
\code{regress} fits a linear model (multiple regression, ANOVA) or
a logistic regression model, depending on the response variable.
}
\details{
The function is a wrapper for \code{\link{lm}} and \code{\link{glm}}
functions. If the response variable has two values (either numeric, factor,
or character), a logistic regression
model is fit (\code{glm} with \code{family=binomial}).
Otherwise a linear model (\code{lm}) is returned.
}
\examples{
# multiple linear regression
fit <- regress(mpg ~ hp + wt, mtcars)
fit

# logistic regression
fit <- regress(am ~ hp + wt, mtcars)
fit
}
\seealso{
\code{\link{lm}}, \code{\link{glm}}
}
