% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relimp.R
\name{relimp}
\alias{relimp}
\title{Variable Importance}
\usage{
relimp(x)
}
\arguments{
\item{x}{an object of class \code{lm} or \code{glm}.}
}
\value{
a ggplot2 graph
}
\description{
The function \code{relimp} uses Dominance Analysis (DA) to rank
order the variables in a linear or logistic regression in terms
of relative importance.
}
\details{
The function is a wrapper for the \link[dominanceanalysis]{dominanceAnalysis}
function in the \code{dominanceanalyis} package. For linear regression
DA analysis uses change in R-squared to assess the average contribution of
each variable to the set all 1, 2, 3, ..., p variable models. For logistic
regression, Estrella's Pseudo R-squared is used (see reference).
}
\examples{
data(mtcars)

# multiple linear regression
fit1 <- lm(mpg ~ ., mtcars)
relimp(fit1)

# logistic regression
mtcars$am <- factor(mtcars$am)
fit2 <- glm(am ~ mpg + cyl +disp + drat + carb, family=binomial, mtcars)
relimp(fit2)
}
\references{
Estrella, A. (1998). A new measure of fit for equations with
dichotomous dependent variables. Journal of Business & Economic
Statistics, 16(2), 198-205.

Azen, R., & Budescu, D. V. (2003). The dominance analysis approach
for comparing predictors in multiple regression. Psychological Methods,
8(2), 129-148.

Azen, R., & Traxel, N. (2009). Using Dominance Analysis to Determine
Predictor Importance in Logistic Regression.
Journal of Educational and Behavioral Statistics, 34(3), 319-347.
}
\seealso{
\link{lm}, \link{glm}
}
