% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lreg.R
\name{summary.lreg}
\alias{summary.lreg}
\title{Summarise an 'lreg' object}
\usage{
\method{summary}{lreg}(x, digits = 4)
}
\arguments{
\item{x}{an object of class "lreg", a result of a call to \link[qacReg]{lreg} (which is a wrapper for \link[base]{glm})}

\item{digits}{number of significant digits to print}
}
\value{
The function \strong{summary.lreg} computes and prints summary statistics and results for a fitted logistic regression model with measures
of robustness including:
\describe{
\item{\strong{Goodness of Fit Measures}}{includes robust goodness of fit measures such as Stukel's test, Psuedo R-squared, and Akaike Information Criterion}
\item{\strong{Anova Table (Type 3 Tests)}}{includes a Type-III Analysis of Variance Table (outputted from the \link[car]{Anova} function in the \strong{car} package)}
\item{\strong{Regression Coefficients}}{includes Coefficients, Standard Errors, t-values, and p-values with significan codes}
\item{\strong{Odds Ratios}}{includes odd's ratios (i.e. exponentiated coefficients) with 95\% confidence intervals}
}
Additionally, summary.lreg also returns tibbles outputted using \link[broom]{glance.lm} and \link[broom]{augment.lm} functions from the \strong{broom} package.
}
\description{
This function serves as a summary method for class "lreg".
}
\examples{
mtcars$am <- factor(mtcars$am)
fit <- lreg(am ~ hp + wt + mpg, mtcars)
summary(fit)
}
