% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relimp.R
\name{varimp}
\alias{varimp}
\title{Variable Importance}
\usage{
varimp(x)
}
\arguments{
\item{x}{an object of class \code{mreg} or \code{lreg}.}
}
\value{
a ggplot2 graph
}
\description{
The function \code{varimp} uses Dominance Analysis (DA) to rank
order the variables in a linear or logistic regression in terms
of relative importance.
}
\details{
The function is a wrapper for the \link[dominanceanalysis]{dominanceAnalysis}
function in the \code{dominanceanalyis} package. For linear regression
DA analysis uses change in R-squared to assess the average contribution of
each variable to the set all 1, 2, 3, ..., p variable models. For logistic
regression, Estrella's Pseudo R-squared is used (see reference).
}
\note{
Estrella, A. (1998). A new measure of fit for equations with
dichotomous dependent variables. Journal of Business & Economic
Statistics, 16(2), 198-205.
}
\examples{
data(mtcars)

# multiple linear regression
fit1 <- mreg(mpg ~ ., mtcars)
varimp(fit1)

# logistic regression
mtcars$am <- factor(mtcars$am)
fit2 <- lreg(am ~ mpg + cyl +disp + drat + carb, mtcars)
varimp(fit2)
}
\seealso{
\link{mreg}, \link{lreg}
}
