% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.lreg.R
\name{diagnostics.lreg}
\alias{diagnostics.lreg}
\title{Diagnostics for Logistic Regression}
\usage{
\method{diagnostics}{lreg}(x)
}
\arguments{
\item{x}{an object of class \code{lreg}}
}
\description{
Provides diagnostics tests and graphs for a logistic regression model
}
\details{

}
\examples{
mtcars$am <- factor(mtcars$am)
fit <- lreg(am ~ hp + wt + mpg, mtcars)
diagnostics(fit)
}
