% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.mreg.R
\name{diagnostics.mreg}
\alias{diagnostics.mreg}
\title{Diagnostics for Multiple Regression}
\usage{
\method{diagnostics}{mreg}(x, output = "brief")
}
\arguments{
\item{x}{an object of class \code{c("mreg", "lm")}}

\item{output}{A parameter with two levels which indicates
whether to display brief diagnostics or
extended (additional) diagnostics.}
}
\description{
Provides diagnostics tests and graphs for a linear model
}
\details{
Provides diagnostics including a graph
for evaluating normality (\code{\link[car]{qqplot}}),
, a graph for evaluating linearity (\code{\link[car]{crPlots}}),
, a graph for evaluating homoscedasticity (\code{\link[car]{spreadLevelPlot}})
as well as a test for heteroscedasticity (\code{\link[car]{ncvTest}})
with a suggested power transformation when it is found.
In addition, if \code{output = "extended"} it provides
a test of multicollinearity (\code{\link[car]{vif}}),
a test for identifying outliers (\code{\link[car]{outlierTest}}),
and a graph for evaluating influential observations
(\code{\link[car]{influencePlot}}).

If the plots do not look right, you may need to clear your plots
by clicking the broom in the plots window in the bottom right corner
of Rstudio

If " Error in plot.new : figure margins too large" is returned,
you must expand the plots window in the bottom right pane of
Rstudio. The error is because this pane is not large enough for the plot,
so by making the plot area larger the function will work.
}
\examples{
mtcars$am <- factor(mtcars$am)
fit <- mreg(mpg ~ wt + am + disp + hp, mtcars)
diagnostics(fit)
diagnostics(fit, output = "extended")
}
\seealso{
\code{\link[qacReg]{diagnostics}}, \code{\link[car]{vif}},
\code{\link[car]{qqplot}}, \code{\link[car]{outlierTest}}, \code{\link[car]{influencePlot}},
\code{\link[car]{crPlots}}, \code{\link[car]{spreadLevelPlot}}, \code{\link[car]{ncvTest}}
}
