% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics.mreg}
\alias{diagnostics.mreg}
\title{Diagnostics for Multiple Regressions}
\usage{
\method{diagnostics}{mreg}(x, output = "brief")
}
\arguments{
\item{x}{an object of class \code{c("mreg", "lm")}}

\item{output}{A parameter with two levels which indicates
whether to display brief diagnostics or
extended (additional) diagnostics.}
}
\description{
Provides diagnostics tests and graphs for a linear model
}
\details{
Provides diagnostics including a graph
for evaluating normality (\code{qqplot}),
, a graph for evaluating linearity (\code{crPlots}),
, a graph for evaluating homoscedasticity
(\code{spreadLevelPlot}), and a test for
evaluating homoscedasticity (\code{ncvTest}).
In addition, if \code{output = "extended"} it provides
a test of multicollinearity (\code{vif}),
a test for identifying outliers (\code{outlierTest}),
and a graph for evaluating influential observations
(\code{influencePlot}).
}
\examples{
fit <- mreg(mpg ~ cyl + wt, mtcars)
diagnostics(fit)
diagnostics(fit, output = "extended")
}
