% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Regression Diagnostics}
\usage{
diagnostics(x, ...)
}
\arguments{
\item{x}{an object of type \code{"lm"} or \code{"glm"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The results of the methods \code{\link{diagnostics.lm}} or \code{\link{diagnostics.glm}}
}
\description{
Regression diagnostics for linear and generalized linear models.
Currently linear and logistic regression models are supported.
}
\details{
The \code{diagnostics} function produces data visualizations
(ggplot2 graphs) for regression diagnostics. The plots depend
on the form of the model (\code{lm} or \code{glm}).
}
\examples{
fit <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
diagnostics(fit)
}
\seealso{
\code{\link{diagnostics.lm}}, \code{\link{diagnostics.glm}}
}
