% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{diagnostics

Run and print regression diagnostics}
\usage{
diagnostics(x, ...)
}
\arguments{
\item{x}{an object of type \code{mreg}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The results of
}
\description{
diagnostics

Run and print regression diagnostics
}
\examples{
fit <- mreg(mpg ~ ., mtcars)
diagnostics(fit)
}
