% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influencePlot.R
\name{gginfluencePlot}
\alias{gginfluencePlot}
\title{Regression Influence Plot}
\usage{
gginfluencePlot(x, n.labels = 2)
}
\arguments{
\item{x}{an object of type \code{"mreg"} or \code{"lm"}.}

\item{n.labels}{integer; the number of points to label.}
}
\description{
This function creates a “bubble” plot of
studentized residuals versus hat values,
with size of the points representing the Cook's distance.
}
\details{
This function is a modification of the \link[car]{influencePlot} function
in the \code{car} package, using \code{ggplot2} rather than
\code{base} graphics.
}
\note{
Vertical reference lines are drawn at twice and three
times the average hat value, horizontal reference lines at
-2, 0, and 2 on the studentized residual scale. The \code{n.label}
parameter controls the number of highest residuals, highest
leverage points, and most influential points to label. For
example \code{n.label=2}, the default, will identify the two
points meeting each criterion, for a maximum of 6 labeled points.
Points meeting more than one criterion are only labeled once.
}
\seealso{
\link[car]{influencePlot}
}
