% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggliftPlot.R
\name{ggliftPlot}
\alias{ggliftPlot}
\title{Lift and Gain Charts}
\usage{
ggliftPlot(actual, prob, positive = "yes")
}
\arguments{
\item{actual}{actual class.}

\item{prob}{predicted probability of target class.}

\item{positive}{which class is the target class?}
}
\value{
a data frame with lift and gain values for each decile.
}
\description{
Create lift and gain charts for binary
classification problems
}
\examples{
# logistic regression example
fit <- glm(caesarian ~ ., family=binomial, data = caesarian)
pred <- predict(fit, newdata=caesarian, type="response")
results <- ggliftPlot(caesarian$caesarian, pred, positive="yes")
print(results)
}
