% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrocPlot.R
\name{ggrocPlot}
\alias{ggrocPlot}
\title{Receiver Operating Characteristic curve}
\usage{
ggrocPlot(actual, prob, positive, n.cuts = 20, labelsize = 3, digits = 2)
}
\arguments{
\item{actual}{actual class.}

\item{prob}{predicted probability for the target class}

\item{positive}{label for the target class}

\item{n.cuts}{number of probability cut-points to plot}

\item{labelsize}{size of cutpoint labels}

\item{digits}{number of decimal digits in the cutpoint labels}
}
\value{
a ggplot2 graph
}
\description{
Plot a receiver operating characteristic curve for a binary predictive model
}
\examples{
# logistic regression
fit <- glm(caesarian ~ ., family=binomial, caesarian)
prob <- predict(fit, newdata=caesarian, type="response")
ggrocPlot(caesarian$caesarian, prob, positive="yes")
}
