% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspreadlevel.R
\name{ggspreadLevelPlot}
\alias{ggspreadLevelPlot}
\title{Spread-Level Plot}
\usage{
ggspreadLevelPlot(x, alpha = 0.4, n.labels = 3, span = 0.75)
}
\arguments{
\item{x}{an object of type \code{"mreg"} or \code{"lm"}.}

\item{alpha}{numeric; degree of transparency for points (0 to 1, default=0.4).}

\item{n.labels}{integer; the number of largest residuals to label
(default=3).}

\item{span}{numeric; smoothing parameter for loess fit line (default=0.75)}
}
\value{
a \code{ggplot2} graph
}
\description{
Creates a plot for examining the possible dependence of
spread on level using the studentized residuals from a linear
model.
}
\details{
This function is a modification of the \link[car]{spreadLevelPlot} function
in the \code{car} package, using \code{ggplot2} rather than
\code{base} graphics.
}
\note{
The graph plots the log of the fitted values against the log
of the absolute studentized residuals. A robust linear fit line
and a loess fit line are also plotted.
}
\examples{
mtcars$am <- factor(mtcars$am)
fit <- mreg(mpg ~ wt + am + disp + hp, mtcars)
ggspreadLevelPlot(fit)
}
\seealso{
\link[car]{spreadLevelPlot}, \link[MASS]{rlm}
}
