% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvif.R
\name{ggvif}
\alias{ggvif}
\title{Variance Inflation Plot}
\usage{
ggvif(x)
}
\arguments{
\item{x}{an object of type \code{"mreg"} or \code{"lm"}.}
}
\value{
a \code{ggplot2 graph}
}
\description{
Plots the variance inflations factors (VIF) for a linear
model.
}
\details{
This function plots results from the \link[car]{vif} function from
the \code{car} package. VIFs larger than 5 are highlighted.

If all terms in an unweighted linear model have 1 df, then the usual variance-inflation factors are calculated.
If any terms in an unweighted linear model have more than 1 df,
then generalized variance-inflation factors (Fox and Monette, 1992)
are calculated.
}
\examples{
mtcars$am <- factor(mtcars$am)
fit <- mreg(mpg ~ wt + am + disp + hp, mtcars)
ggvif(fit)
}
\references{
Fox, J. and Monette, G. (1992) Generalized collinearity diagnostics.
\emph{JASA}, \emph{87}, 178–183.
}
\seealso{
\link[car]{vif}
}
