% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.lm.R
\name{info.lm}
\alias{info.lm}
\title{Summarise an 'lm' object}
\usage{
\method{info}{lm}(x)
}
\arguments{
\item{x}{an object of class "lm"}
}
\value{
a 6-component list
\describe{
\item{fit.indices}{includes R-squared, Adjusted R-squared, Akaike's Information Criterion, Root Mean Squared Error, Mean Absolute Error}
\item{Ftest}{includes F-statistic, Degrees of Freedom, p-value}
\item{anova.table}{includes a Type-III Analysis of Variance Table (outputted from the \link[car]{Anova()} function in the \strong{car} package)}
\item{coefficient.table}{includes a table of Coefficients, Standardised Coefficients, Standard Errors, t-values, p-values, significance stars}
\item{N}{sample size}
}
}
\description{
Summarize an object of class \code{"lm"}.
}
\details{
The function \code{info.lm} produces a comprehensive
summary for a linear model fitted with \link{lm} function. In addition
to the standard output from the \link{summary.lm} function, the output includes
the sample size (N), Akaike's Information Criterion (AIC), Mean Absolute Error (MAE), ANOVA table (type III SS), and
standardized regression coefficients (B*).
}
\examples{
fit <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
info(fit)
}
