% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Summary information for lm or glm models}
\usage{
info(x)
}
\arguments{
\item{x}{an object of type \code{lm} or \code{glm}.}
}
\value{
The results of the methods \code{\link{info.lm}} or \code{\link{info.glm}}
}
\description{
Detailed summary information of a linear or logistic
regression model
}
\examples{
fit <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
info(fit)
}
