% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mreg.R
\name{mreg}
\alias{mreg}
\title{Multiple Regression}
\usage{
mreg(formula, data, ...)
}
\arguments{
\item{formula}{an object of class formula.}

\item{data}{a data frame.}

\item{...}{arguments passed to the \code{lm} function.}
}
\value{
an object of type \code{c("mreg", "lm")}
}
\description{
Fit a linear model of the form
\eqn{y = \beta0 + \beta1*X1 + \beta2*X2 + ...+ \betak*Xk}.
}
\details{
\code{mreg} is a wrapper for the \code{\link{lm}} function.
}
\examples{
fit <- mreg(mpg ~ hp + wt + am, mtcars)
}
