% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.glm.R
\name{performance.glm}
\alias{performance.glm}
\title{Model Performance for Logistic Regression}
\usage{
\method{performance}{glm}(x, data, digits = 4, p = 0.5, plot = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{"glm"}.}

\item{data}{a data frame. If no data frame is provided, the model training
data is used.}

\item{digits}{integer; number of digits to print (default=4).}

\item{p}{numeric; probability cutoff for classifying cases (default = 0.5)}

\item{plot}{logical; If \code{TRUE} the results are plotted.}

\item{...}{not currently used}
}
\value{
The results of the methods \code{\link{performance.glm}}
}
\description{
Produce indices of model performance for logistic regression
}
\details{
The model in the \code{glm} object is used to predict the
the predicted probability of being a "positive" case for each observation
in the data frame. If a data frame is not specified, the model is
applied to the training data (data frame contained in the model component
of the \code{glm} object).
If the predicted probability is greater then \code{p}, the case is
classified as "positive", "not positive" otherwise. The actual and
predicted classifications are compared to calculate performance
metrics.

The function returns general information
about the model, a confusion matrix and statistics generated by
the \code{\link[caret]{confusionMatrix}} function from the \code{caret}
package, and a graph visualizing the results.
}
\examples{
# performance on training sample
fit <- regress(caesarian ~ ., data = caesarian)
performance(fit, plot=TRUE)
}
\seealso{
\code{\link[caret]{confusionMatrix}}, \code{\link{roc_plot}},
\code{\link{lift_plot}}.
}
