% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.lm.R
\name{performance.lm}
\alias{performance.lm}
\title{Model Performance for Linear Models}
\usage{
\method{performance}{lm}(x, data, digits = 4, N = 20, ...)
}
\arguments{
\item{x}{an object of class \code{"lm"}.}

\item{data}{a data frame.}

\item{digits}{integer; number of digits to print (default=4).}

\item{N}{integer; number of bootstrap replications (default=20).}

\item{...}{not currently used}
}
\value{
The results of the methods \code{\link{performance.lm}}
}
\description{
Model Performance for Linear Models
}
\details{
Produce indices of model performance for linear models
}
\examples{
# performance on training sample
fit <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
performance(fit)
}
