% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Model Performance

Produce indices of model performance}
\usage{
performance(x, ...)
}
\arguments{
\item{x}{an object of type \code{"lm"} or \code{"glm"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The results of the methods \code{\link{performance.lm}} or
\code{\link{performance.glm}}
}
\description{
Model Performance

Produce indices of model performance
}
\examples{
# multiple regression
fit <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
performance(fit)

# logistic regression
# fit <- lm(am ~ hp + wt, mtcars)
# performance(fit)
}
