% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mreg.R
\name{plot.mreg}
\alias{plot.mreg}
\title{Plotting multiple regression}
\usage{
\method{plot}{mreg}(x, points = FALSE)
}
\arguments{
\item{x}{an object of type \code{c("mreg", "lm")}}

\item{points}{points is TRUE if data points are to be shown, default is FALSE}
}
\value{
a plot of class \code{ggplot}
}
\description{
The plot function provides a single plot that shows separate graphs of the relationship
between each predictor and the outcome, controlling for the other predictors from the
mreg function
}
\examples{
fit <- mreg(mpg ~ hp + wt + am, mtcars)
plot(fit)
plot(fit, points=TRUE)
}
