% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mreg.R
\name{plot.mreg}
\alias{plot.mreg}
\title{Plotting multiple regression}
\usage{
\method{plot}{mreg}(
  x,
  points = TRUE,
  dot.size = 1,
  dot.alpha = 0.5,
  ci = TRUE,
  page = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of type \code{c("mreg", "lm")}}

\item{points}{points is TRUE if data points are to be shown, default is FALSE}

\item{dot.size}{size of data points, default is 1}

\item{dot.alpha}{transparency of data points, default is 0.5}

\item{ci}{if confidence intervals are to be shown, default is TRUE}

\item{page}{logical. If \code{TRUE}, each graph will appear separately.}

\item{...}{additional arguments to be added to adjust graphic}
}
\value{
a plot of class \code{ggplot2}
}
\description{
The plot function provides a single plot that shows separate graphs of the relationship
between each predictor and the outcome, controlling for the other predictors from the
mreg function
}
\details{
The \code{plot.mreg} function is uses ggeffects package to plot the results of the mreg function from the
qacReg package. The ggeffects package aims to easily calculate marginal effects and adjusted predictions
at representative values of covariates from statistical models, i.e. predictions generated by a model when
one holds the controlled dependent variable(s) constant and varies the independent variable(s).
To see more details and gain better understanding of the ggeffects package see
\link[https://strengejacke.github.io/ggeffects/]{here}
}
\examples{
mtcars$am <- factor(mtcars$am)
fit <- mreg(mpg ~ hp + wt + am, mtcars)
plot(fit)
plot(fit, points=TRUE)
}
\seealso{
\link[ggeffects]{ggeffect}
\link[ggplot2]{ggplot}
}
