% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\title{Marginal Effects Plots}
\usage{
plots(
  x,
  points = FALSE,
  size = 1,
  alpha = 0.2,
  jitter = 0.03,
  ci = TRUE,
  page = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"lm"} or \code{"glm"}}

\item{points}{logical; If \code{TRUE}, raw data points are plotted (default = \code{FALSE})}

\item{size}{size of data points (default = 1)}

\item{alpha}{transparency of data points (from 0 to 1; default is 0.2)}

\item{jitter}{amount of jitter for points (from 0 to 1; default = .03)}

\item{ci}{logical; If \code{TRUE}, display 95\% confidence intervals (default = \code{TRUE})}

\item{page}{logical; If \code{TRUE}, each graph will be printed separately.}

\item{...}{additional arguments to be passed to
\code{\link[ggeffects]{plot.ggeffects}}.}
}
\value{
a list of \code{ggplot2} graphs.
}
\description{
The \code{plots} function generates one or more graphs displaying the relationship
between each predictor and the outcome, controlling for the other predictors
in the model.
}
\details{
The \code{plots} function uses the \code{\link[ggeffects]{ggeffect}} and
\code{\link[ggeffects]{plot.ggeffects}} functions from the  \code{ggeffects} package
to plot marginal effects for a linear or generalized linear model.

A plot is created for each predictor variable (called the "focal" variable).
Predicted response values are calculated by varying the values of
the focal variable while keeping the non-focal variables
(the other predictor variables) constant.
Non-focal numeric variables are set to their means. If factors are
present in the model, predicted
values are calculated for each combination of
non-focal categorical variable levels and averaged based on category size. See the
\href{ggeffects homepage}{https://strengejacke.github.io/ggeffects/index.html}
for details.

If \code{page = TRUE},
the individual graphs are combined into a single plot via the
the \code{\link[patchwork]{wrap_plots}} function in the patchwork package.
Otherwise, individual plots are printed.
}
\section{Limitations}{

The \code{plots} function can not handle models that include
\code{as.factor} or \code{factor} in their formulas. Create the modified
variables prior to fitting the model
(e.g., \code{mtcars$cyl <- factor(mtcars$cyl)}).
Additionally, the function can not handle models that include interactions. In this
case, consider using the \code{ggeffects} package with grouping and
faceting.
}

\examples{
#######################
# Multiple regression #
#######################
fit <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
plots(fit)

#######################
# Logistic regression #
#######################
fit2 <- glm(caesarian ~ age + bp + delivery.time, family = binomial, data = caesarian)
plots(fit2)
}
\seealso{
\link[ggeffects]{ggeffect}, \link[ggeffects]{plot.ggeffects}
}
