% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\title{Plotting for linear and logistic regression models}
\usage{
plots(
  x,
  points = TRUE,
  size = 1,
  alpha = 0.5,
  jitter = 0.03,
  ci = TRUE,
  page = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"lm"} or \code{"glm"}}

\item{points}{points is TRUE if data points are to be shown, default is FALSE}

\item{size}{size of data points, default is 1}

\item{alpha}{transparency of data points, default is 0.5}

\item{jitter}{ammount of point jitter (from 0 to 1)}

\item{ci}{if confidence intervals are to be shown, default is TRUE}

\item{page}{logical. If \code{TRUE}, each graph will appear separately.}

\item{...}{additional arguments to be added to adjust graphic}
}
\value{
a plot of class \code{ggplot2} if \code{page=FALSE}. a list
of separate \code{ggplot2} graphs, if \code{page=TRUE} and there is
more than one explanatory variable.
}
\description{
The \code{plots} function provides a single plot that shows separate graphs of the relationship
between each predictor and the outcome, controlling for the other predictors in the model.
}
\details{
The \code{plots} function is uses ggeffects package to plot the results of the lm or glm functions
The ggeffects package aims to easily calculate marginal effects and adjusted predictions
at representative values of covariates from statistical models, i.e. predictions generated by a model when
one holds the controlled dependent variable(s) constant and varies the independent variable(s).
To see more details and gain better understanding of the ggeffects package see
\href{https://strengejacke.github.io/ggeffects/}{here}.
}
\examples{
fit1 <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
plots(fit1)

mtcars$am <- factor(mtcars$am)
fit2 <- glm(am ~ hp + wt + mpg, family=binomial, mtcars)
plots(fit2)
}
\seealso{
\link[ggeffects]{ggeffect}, \link[ggplot2]{ggplot}
}
