% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.info.glm.R
\name{print.info.glm}
\alias{print.info.glm}
\title{Print an "info.glm" Object}
\usage{
\method{print}{info.glm}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{"info.glm"} produced by the
\code{"info.glm"} function.}

\item{digits}{number of significant digits to print.}

\item{...}{parameters passed to print.}
}
\description{
\code{print} method for class \code{"info.glm"}.
}
\examples{
fit <- glm(caesarian ~ age + bp + delivery.time, family = binomial, data = caesarian)
sfit <- info(fit)
print(sfit, digits=5)
}
