% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{summary.mreg}
\alias{summary.mreg}
\title{Summarise an 'mreg' object}
\usage{
\method{summary}{mreg}(x)
}
\arguments{
\item{x}{an object of class 'mreg', presumably an output from the mreg() function.}
}
\value{
a list containing summary statistics and results
}
\description{
This function serves as a summary method for class 'mreg'.
It computes summary statistics and results for a fitted linear model.
}
\details{

}
\examples{
fit.mreg <- mreg(mpg ~ cyl + hp + wt, mtcars)
summary.mreg(fit.mreg)
}
