% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mreg.R
\name{summary.mreg}
\alias{summary.mreg}
\title{Summarise an 'mreg' object}
\usage{
\method{summary}{mreg}(x, k = 5)
}
\arguments{
\item{x}{an object of class "mreg", a result
of a call to \link{mreg} (which is a wrapper
for \link{lm})}

\item{k}{number of folds for cross validated performance}
}
\value{
The function \strong{summary.mreg} computes summary statistics and results for a fitted linear model with measures
of robustness including:
\describe{
\item{fit.indices}{includes R-squared, Adjusted R-squared, Akaike's Information Criterion, Root Mean Squared Error}
\item{cv.indices}{includes a more reliable and accurate measure of R-squared, Root Mean Squared Error}
\item{Ftest}{includes F-statistic, Degrees of Freedom, p-value}
\item{anova.table}{includes a Type-III Analysis of Variance Table (outputted from the \link[car]{Anova()} function in the \strong{car} package)}
\item{coefficient.table}{includes a table of Coefficients, Standardised Coefficients, Standard Errors, t-values, p-values, Significance codes}
\item{k}{number of folds for cross-validation}
}
}
\description{
Summarize an object of class \code{"mreg"}.
}
\examples{
mtcars$am <- factor(mtcars$am)
fit <- mreg(mpg ~ hp + wt + am, mtcars)
summary(fit, k=2)
}
